requirements.txt
aiogram>=3.20.0
psycopg2-binary>=2.9.5
